Move Rollup task to separate file
